[3.14] GH-133600: Backport file reorg of `Tools/wasm/wasi`
