+++ Misc/NEWS.d/next/Tools-Demos/2025-06-25-10-36-22.gh-issue-133600.bkdgHC.rst
@@ -1,0 +1,4 @@
+Backport file reorganization for Tools/wasm/wasi .
+
+This should make backporting future code changes easier. It also simplifies
+instructions around how to do WASI builds in the devguide.
